@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -27,12 +27,14 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  google_fonts:
-  http: ^0.12.0+4
-  story_view: ^0.12.5
-  flutter_map: ^0.10.1+1
-  timeline_tile: ^1.0.0
+  cupertino_icons: ^1.0.6
+  google_fonts: ^6.1.0
+  http: ^1.1.0
+  # story_view: ^0.15.0  # Commented out due to compatibility issues
+  flutter_map: ^6.1.0
+  latlong2: ^0.9.1
+  timeline_tile: ^2.0.0
+  provider: ^6.1.1
 
 dev_dependencies:
   flutter_test:
